requirements.txt
altgraph==0.17.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astroid==2.9.3
asttokens==2.1.0
attrs==22.1.0
auto-py-to-exe==2.18.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
colorzero==2.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.13.0
django-unused-media==0.2.2
djangorestframework==3.14.0
djlint==0.7.6
easygui==0.98.3
Eel==0.12.4
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.6.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
gpiozero==1.6.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
iso8601==1.0.2
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.10
jsonschema==4.17.0
jupyter-server==1.23.1
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==2.0.4
MouseInfo==0.1.3
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
netaddr==0.8.0
nose==1.3.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.2
opencv-python==4.5.5.64
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pbr==5.8.1
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.0.1
pip-api==0.0.30
pipenv==2022.7.4
platformdirs==2.5.1
playsound==1.2.2
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.0
pure-eval==0.2.2
py-solc-x==1.1.1
py4j==0.10.9.5
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.13.0
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
pylint==2.12.2
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.19.2
PyScreeze==0.1.28
PySide6==6.2.3
PySignal==1.1.1
pyspark==3.3.1
pytesseract==0.3.9
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.6
PyVirusShare==1.0.3
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
semantic-version==2.9.0
Send2Trash==1.8.0
serial==0.0.97
shiboken6==6.2.3
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.0
stevedore==3.5.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.9.2
tornado==6.2
tqdm==4.63.1
traitlets==5.5.0
tzdata==2022.1
unicorn==1.0.3
urllib3==1.26.9
varint==1.0.2
virtualenv==20.14.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
waitress==2.1.1
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
whichcraft==0.6.1
whitenoise==6.1.0
WMI==1.5.1
wrapt==1.13.3
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
